inline needed data as a function, get rid of additional workflow steps
